avoid replacing playback.html when install is executed 2nd time

+++ bbb-mp4-install.sh
@@ -14,16 +14,31 @@ mkdir "$mp4_dir"
 mkdir "$download_dir"
 
 echo "Updating post_publish.rb"
-mv /usr/local/bigbluebutton/core/scripts/post_publish/post_publish.rb /usr/local/bigbluebutton/core/scripts/post_publish/post_publish.rb.default
-cp post_publish.rb /usr/local/bigbluebutton/core/scripts/post_publish/
+if [ ! -f "/usr/local/bigbluebutton/core/scripts/post_publish/post_publish.rb.default" ]; then
+  echo "post_publish.rb.default doesn't exist. Proceed with replacing.";
+  mv /usr/local/bigbluebutton/core/scripts/post_publish/post_publish.rb /usr/local/bigbluebutton/core/scripts/post_publish/post_publish.rb.default
+  cp post_publish.rb /usr/local/bigbluebutton/core/scripts/post_publish/    
+else
+  echo "post_publish.rb.default exists. Skipping replacing.";
+fi
 
 echo "Updating playback.html"
-# playback_default.html is used by bbb-mp4 for recording
-cp /var/bigbluebutton/playback/presentation/2.0/playback.html /var/bigbluebutton/playback/presentation/2.0/playback_default.html
-# playback.html.default is for refernce, in case you want to restore default BBB playback 
-mv /var/bigbluebutton/playback/presentation/2.0/playback.html /var/bigbluebutton/playback/presentation/2.0/playback.html.default
-# the new playback.html being copied will direct users to /recording/<meeting-id>.mp4
-cp playback.html /var/bigbluebutton/playback/presentation/2.0/playback.html
+if [ ! -f "/var/bigbluebutton/playback/presentation/2.0/playback.html.default.html" ]; then
+  
+  echo "playback.html.default doesn't exist. Proceed with replacing.";
+  
+  # playback_default.html is used by bbb-mp4 for recording
+  cp /var/bigbluebutton/playback/presentation/2.0/playback.html /var/bigbluebutton/playback/presentation/2.0/playback_default.html
+  
+  # playback.html.default is for refernce, in case you want to restore default BBB playback 
+  mv /var/bigbluebutton/playback/presentation/2.0/playback.html /var/bigbluebutton/playback/presentation/2.0/playback.html.default
+  
+  # the new playback.html being copied will direct users to /recording/<meeting-id>.mp4
+  cp playback.html /var/bigbluebutton/playback/presentation/2.0/playback.html
+else
+  echo "playback.html.default exists. Skipping replacing.";
+fi
+
 
 echo "Installing xvfb"
 apt-get -y update
